:bug: Reduce intersection observer threshold (#4824)
* :bug: Reduce intersection observer threshold

When this value is 1, scrolling down the page (sometimes?) triggers the value to change - probably something changes invisibly. Setting it to anything less than 1 seems to cure the issue, and there are no longer unwanted triggers of changes. 1% also seems to make more sense here than 100% - we want the webcam to load whenever even a tiny amount of it is visible.

* ♻️ Change visibility observer threshold to 1%

* ♻️ Use log.debug instead of console.debug

---------

Co-authored-by: Gina Häußge <gina@octoprint.org>