✨ Add refresh button for webcam feed
Will call `onWebcamRefresh` on the bound viewmodel of the currently
active webcam. It is up to the implementing webcam view model
to execute the actual refresh.

Only the mjpg stream type of the classicwebcam currently supports
this, for the other stream types it is a no-op.

Note that the button will be displayed on mouseover regardless of
whether it will actually be capable of triggering a refresh or not.

Also adjusted the fade-in animation of the keycontrol overlay to
match the refresh button and webcam selector.

Implements #4617