🔒️ Use sandboxed jinja env for gcode scripts
Without this patch, an attacker with admin rights could configure
a specially crafted GCODE Script to allow for code execution during
rendering of that script. An attacker might use this to extract data
managed by OctoPrint, or manipulate data managed by OctoPrint,
as well as execute arbitrary commands with the rights of the
OctoPrint process on the server system.

Fixes a reported security issue, CVE pending.