🥅 virtual: recalculate wait timeout after gcode eval
Otherwise we might run into a race condition in case of long running
commands.

For an example see #4907:

A `G4 S54` triggered a `wait` response, since right after sending the
`ok` the host didn't send a new incoming line fast enough to prevent
that from happening. As a result, combined with several simulated
transmition errors right after, things get so much out of balance that
at some point the incoming queue runs full and an exception gets
triggered, killing the connection with a simulated SerialTimeout.

This fix should close the core cause of #4907