🐛 Fix double response on POST/PUT/etc
Not setting the _finished flag to True caused the UploadStorage-
FallbackHandler to generate an additional empty text/html response
for every POST/PUT/etc request. If that arrived inside a keep-alive
connection followed by additional requests, it caused the browser
to consider the text/html response to be the response to the follow-up
request, causing all kinds of mayhem.