🐛 virtual: remove wrong assertion
Was actually wrong to have that here, but thankfully it was
implemented wrong as well, and here two wrongs made
one right.

To match existing firmware out there, we should be sending
SD status auto reports with progress information whenever
a print from a selected SD file has been started. If a such a print
is cancelled via OctoPrint, all we can do is send an M25 to pause
it and just never resume it. That means that the SD status will
continue to report the current file position.

See #4924