✨ Add authenticated reverse proxy test w/ more data
If accessed through with `?authenticated` parameter, the page will
now also render the relevant request headers and output the
configured trusted proxies, after confirming ADMIN permissions.

That way admins get easier access to information needed to
debug semi-broken reverse proxy setups, without leaking such
information to everyone.

A general access restriction of the reverse proxy test page is not
possible as if logins are broken thanks to a wrong config, users
would then not be able to access it to figure out why. This is
a compromise to solve this dilemma.