🥅 Fix a potential race condition
I observed several build failures yesterday due to an attempt to
create the data folder failing because it already existed.

I don't know if this was due to a race condition between check
and creation or some other error. However, the code indeed was
not thread safe, now it should be, so IF that was indeed the case
this should not happen again.