👷 Move cleanup workflow to earlier in the night
Let's see if this maybe already helps with the rate limits.

Switching everything to REST would work fine for the closing of
incomplete issues, but for the locking of old issues and PRs it
proves to be a challenge given that on the REST API we can't limit
things to only "last update older than 1y and not yet closed" easily
and thus would have to always paginate through all results, which
sounds like a bad idea.