⏪ Revert to netifaces
While netifaces2 as used in 1.10.0 works well, it is sadly causing
some build issues in the field. In the interest of giving as many
people as possible access to any bug and especially security
fixes, we are thus reverting to the (unmaintained) netifaces
for now and keeping an eye on the wheel availability und
compatibility of netifaces2 for a future repeated rollout.