config: change trustedDownstream to trustedProxies + flag
As outlined in #5036, trusted proxy configuration was a bit confusing up
until now. localhost wasn't added automatically in case additional proxies
were configured, leading to hard to debug issues in the field.

Additionally, there was some confusion with regards to trustedDownstream
vs trustedUpstream, the latter being a typo that crept into the code.

The setting has now been renamed to trustedProxies. An additional flag
trustLocalhostProxies defaulting to True will ensure that localhost 127.0.0.0/8
and ::1 will always be trusted, if set. trustedDownstream will be migrated to
those two config options - any reverse proxy that isn't 127.0.0.0/8, ::1 or
127.0.0.1 will be put into trustedProxies, trustLocalhostProxies will be set
depending on whether any of the localhost addresses previously was set
in trustedDownstream or not.

Unit tests have been added to test this behaviour, and the documentation
for config.yaml has been adjusted as well.

Closes #5036.