feat: Python EOL tracking & upgrade heads-up
OctoPrint will now check the current version against a list of known EOL
versions with optional end of support version for OctoPrint itself, and
warn the user if they are running on an EOL version on every page load.

TODO: Should we add some kind of cookie to only show a warning
every week or so, at least for future EOL versions?