fix[ux]: Display python eol popup only once per 30d
Uses a cookie with a 30d max-age to ensure that the popup doesn't get
shown on every UI reload.

The cookie value is set to the combination of python version and current
EOL date, so if that changes, the popup will be shown again.

A cookie value of "true" will always hide the popup, this is useful for E2E
testing where we don't want the popup from interfering with our automated
clicks.