refactor: get rid of pkg_resources use
It's part of setuptools, no longer available by default on newer installs, and while
we currently still depend on setuptools, we might want to change that sooner or
later.

In any case, now that there are better options to do these things with core and/or
dedicated python tooling, let's use that instead of pkg_resources.