feat: add bundled health_check plugin
Checks the following:
- sufficient file system storage across all configured data folders
- OctoPrint's version not being more than two minor versions behind
- Python version not being EOL or close to EOL

Additional health checks can be added by plugins through the
octoprint.plugin.health_check.get_additional_checks hook.

An update of the PiSupportPlugin is forthcoming that adds the
results of the throttle check, model support and default password use
to the set of health checks.

This is one step towards reducing the amount of outdated
environments used by people and giving them early heads-ups
that they need to update to not be left behind.