feat[upmgr]: Add a bundled Upload Manager plugin (#5062)
Bundles a new plugin "Upload Manager", which is a plugin derived from the long abandoned Filemanager plugin authored by @Salandora.

This Upload Manager can be access through a new button in the file list header and allows various bulk operations to more easily manage a growing file collection.

With the availability of this Upload Manager, confirmation dialogs for file deletion have been added, as requested in #833. With the manager in place and thus a bulk deletion that only prompts for confirmation once, this should not interrupt any existing management workflows too much.

Added functionality vs the Filemanager plugin:
- progress information for bulk operations
- download of multiple files or folders as zip files
- multi select with shift/ctrl-click
- collision detection on file rename
- various UX improvements

Implements #833