refactor: make compatible to Python 3.13
The primary issue was a whole ton of deprecation warnings caused
by both core and third party dependencies. Upgrading the latter
proved to be complicated due to the fixed versions having dropped
Python 3.7 support. So we now have some more python version splits
in the requirements.

A move away from Python 3.7 needs to be done ASAP and is
scheduled for OctoPrint 1.12.0.

The server startup saw some *heavy* refactoring in order to get
rid of an ioloop related deprecation warning, followed by some
reorganization to have things slightly more manageable in the
future.