fix: use first matching cookie instead of last
Normally we shouldn't have more than one cookie of the same name,
but some Firefox instances might disagree [1] and thus we make sure
we at least use the same order on the frontend as on the backend.

This won't solve reading the wrong cookie under all circumstances
(clearing cookies via the devtools causes some fun effects), but
reduces the risk of this happening.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1929881