feat(swu): remove --no-cache-dir from pip command line
The pip bug for which that parameter was originally added
(pypa/pip#3532) has been fixed since pip 19.3. We require pip
versions above 20.3 these days anyhow to operate, so this bug
should always be fixed in whatever we encounter.

Not being able to use pip's cache dir has significant performance
impact when it comes to installing dependencies only available
as source (their wheels need to be rebuilt on every update), so
let's get rid of this now.

Reverses #2543