fix: set login mechanism for api key requests
As it turns out, after switching to a custom identify loader relying
on flask's current_user in #5098, we forgot to also adjust the api key
handler to set the login_mechanism (and last credentials seen
timestamp) as it is being done with the header handlers.

This caused anything but GET, HEAD and OPTIONS request done
with an API key to run through the CSRF check, which they would
obviously fail as expected.

Closes #5100