feat(api): add support for API versioning
The API version to use can be set by the client either through a new header
`X-OctoPrint-Api-Version` or a request parameter `api_version`. The latter should
only be used for testing though.

The JS client and the UI have been adjusted to make use of that to always request
the API version matching the current base version of OctoPrint.