fix(coreui): process job data only once
When selecting a file from the SD card, things like the size
might be missing from the info we get from the backend. In
that case the UI would currently try to fetch the data twice per
second while that file is selected, as it can't detect that we have
more info from the job data than on the file manager.

To solve this, we no longer only check for equality but also for
prefixes. That way, we can match up something like
`sdcard:file.gco:null:null` and `sdcard:file.gco:123:null`, by
stripping the `:null` parts (`sdcard:file.gco` and
`sdcard:file.gco:123`) and then checking for a full prefix incl `:`.

Partial fix for #5121