fix(cli): ensure migrated plugins support setuptools<77
A plugin might still support Python 3.7 or 3.8, yet setuptools>=77 doesn't.
With setuptools>=77 the project.license field needs to be a (SPDX) string,
before that a table. Since we don't want to pin setuptools<77, we therefore
require some solution that will dynamically solve this situation and ensure the
license will always end up in the meta data.

Thus, migrated plugins will get a setup.py after all, and get the project.license
set to dynamic. setup.py then simply sets the license parameter via the license
parameter (which hopefully will continue to work, at least it is documented).

It would of course have been nicer to be able to check the setuptools version
and only do all of this crap if it's <77, but alas, there doesn't seem to be a
solution for that. It is possible to instead of license set license_expression in
newer versions, but that's as of now undocumented and thus cannot be
used.

See also pypa/setuptools#4903 for more details.